+++ app/schema/index.js
@@ -1,0 +1,5 @@
+const Ajv = require('ajv');
+
+const ajv = new Ajv({ allErrors: true });
+
+module.exports = ajv;
